pyproject.toml
[project]
name = "toutiao-agent"
version = "0.1.0"
description = "头条热点自动评论助手"
requires-python = ">=3.10"

dependencies = [
    "click>=8.0.0",
    "playwright>=1.40.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
toutiao-agent = "toutiao_agent.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/toutiao_agent"]
